fix(docs,config): ensure dev env guide is presentable
Styles for admonitions were added and Jekyll plugin files were added to override the admonitions that come with "ExtendedMarkdownFilter". More work can be done on improving these and other filters included by this plugin. Operating system-specific instructions will be needed to be used for this guide in the future. The default styles for that filter don't seem to be sufficient for our purposes.