📄✨：temporarily fixing a google indexing error
This message was received earlier today; we want this a little better for
now; that's all we should need to help w/ the 404 page coming up dupe
content for this URL (Googlebot must be interested). :)

```text
Search Console has identified that your site is affected by 1 Page indexing issue(s). The following issues were found on your site.

Top Issues

    Duplicate, Google chose different canonical than user

We recommend that you fix these issues when possible to enable the best experience and coverage in Google Search.
```

Signed-off-by: DerekNonGeneric <dereknongeneric@open.inf.is>